Support IPv6 on wwan interface
If the wwan interface comes up with an IPv6 gateway, add an IPv6 /64 to
the bridge for the local network, masquerading this range onto the single
wwan address (as with IPv4) to avoid contravening network device limits.

Unfortunately, hosts implementing RFC 6724 address selection will prefer
IPv4 private addresses to IPv6 ULA. To work around this, repurpose
2001:2/64 from the reserved benchmarking range for our local network,
so hosts will use an IPv6 destination when available.